refactor(automations): move schedule trigger into builtin package
