+++ surfsense_backend/app/automations/triggers/builtin/__init__.py
@@ -1,0 +1,5 @@
+"""Built-in trigger types — each in its own subpackage, self-registering at import."""
+
+from __future__ import annotations
+
+from . import event, schedule  # noqa: F401
